deploy: generate service account token signing key
This RSA key allows the ServiceTokensController to mint new
tokens in the form of Kubernetes Secrets for ServiceAccounts.

The apiserver is not yet enforcing authentication, but the
ServiceAccount admission controller is enabled. This ensures
that new pods are wired up with the default ServiceAccount
for a given Namespace if one is not specified.