README: downloading k8s can take a while, so reorder
Downloading all of the k8s containers can take a while so instead of
having the user downlaod and configure kubectl first, lets do that after
they get the machine booting. That way they are distracted for a minute
or two and hopefully the containers can get running in the meantime.