multi-node/aws/artifacts: Add node labels for AWS metadata.
This demonstrates a proposal on how to tag nodes with labels derived from AWS metadata.  It makes use of a small single-shot container dropped into the /etc/kubernetes/metadata directory to fetch the metadata and apply the label.

The container I use can be found in Docker Hub as elevy/aws-node-labels.  CoreOS would copy and host the container for it to be officially integrated.

Note that this is just a stopgap until the https://github.com/kubernetes/kubernetes/pull/13524 proposal is implemented.

Fixes #120.