kube-aws: provision all artifacts via cloudformation
This removes the need for an s3 intermediary for hosting artifacts.

When kube-aws up is invoked, all the artifacts are compressed and packed into the cloudconfig template supplied via the cloudformation API call.

The resulting cloudformation template now contains the entirety of the deployment vector--  no dependencies on what's in an s3 bucket.