Add support for network policy using Calico
This (rather large) PR adds Calico to the Vagrant and AWS installs and
the documentation.

In all cases, Calico is enabled using a feature flag. It's enabled by
default for single host vagrant installs, and disabled by default for
all others.

Calico relies on using CNI networking plugins with Kubernetes, which
currently means that a special hyperkube images (with a .cni suffix) is
required.