kube-aws: better support for AWS credentials file
Using the SharedCredentials codepath allows the user to better
control what AWS credentials are used. Specifically, the ability
to set AWS_PROFILE=<profile> is desireable so that users do not
have to export credentials into environment variables, and are
no longer required to set credentials to the "default" profile
in ~/.aws/credentials.