kube-aws: don't add trailing dot to empty dns names
If you add the trailing dot to an empty hostname, the validation
that hostname isn't blank is worthless.

Basically none of the trailing dot stuff was being tested because it was
being added in ClusterFromFile rather than ClusterFromBytes, which is only
used in the live code. Also, the r53 dummy service was returning the
HostedZone and ResourceRecordSet without the final ., which would be
there in the real return values.