kube-aws: Fix two bugs with subnets & existing VPCs
Bug #1: Before this patch, when specifying a route table ID in the
cluster.yml file, the substitution on line 618 fails because the
RouteTableID is not available in that scope. This patch modifies the
substitution to read from the correct scope.

Bug #2: Before this patch, the validation code that is run during
'kube-aws up' expects ``c.InstanceCIDR`` to be a valid CIDR even when
the Subnets parameter is set instead. This patch modifies the validation
to avoid raising errors in this case.