Propagate aws kms decrypt's exit code in decrypt-tls-assets
decrypt-tls-assets runs `aws kms` in a docker container, and then
pipes the output to `base64`. This does not return a non-zero exit
code when decryption fails, and so the unit is not retried.

With this commit, we run the two commands separately so that
bash -e barfs if the decrypt fails.