Mark sections as optional in deploy-workers.md
k8s/deploy-workers.md: Add "(optional)" to headers of optional sections

This is mainly so that it matches the deploy-master document. Setting up Calico is optional. As a newb, I had to go back to the master instructions and make sure that these parts were optional for the workers as well.