docs: Remove extra instance deployment
If you follow this guide, it will only provision a worker node and
controller node. In order to deploy 2 workers, you'd need to modify
workerCount in the cluster.yaml file before rendering the
CloudFormation template.

This commit corrects the output to what the user sees. Another question
is whether this guide should have users increase that value so their
sample deployment features > 1 worker.